refactor(interpreter): move distribution.hpp into `random` directory
